requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
click==8.1.7
    # via -r requirements.in
pillow==10.2.0
    # via -r requirements.in
